requirements.txt
selenium
python-dotenv
pyautogui
pandas
openpyxl